requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn
Automat==20.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
boto3
botocore
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
click==8.0.1
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-extensions==0.0.67
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
defusedxml==0.7.1
distlib==0.3.2
Django
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==3.0.1
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-database-view==0.3.0
django-db-views==0.1.1
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-plotly-dash==1.6.5
django-webpack-loader==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dnspython==1.16.0
dpd-components==0.1.0
dpd-static-support==0.0.5
ecdsa==0.17.0
EditorConfig==0.12.3
eutils==0.6.0
fair-research-login==0.2.6
filelock==3.0.12
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Compress==1.10.1
future==0.18.2
globus-cli==3.1.4
globus-sdk==3.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jose==1.0.0
jsbeautifier==1.14.0
jsonpickle==2.0.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
minio==7.1.0
mongoengine==0.22.1
more-itertools==8.12.0
mysqlclient==2.0.3
numpy
oauthlib==3.1.1
packaging==21.0
pandas
pipenv==2021.5.29
platformdirs==2.3.0
plotly==5.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pymemcache==3.5.0
pymongo
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
selenium
service-identity==21.1.0
six==1.16.0
social~=0.1.dev3
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.2.1
sqlparse==0.4.1
tenacity==8.0.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.7
Werkzeug==2.0.1
whitenoise==5.3.0
xmltodict==0.13.0
zope.interface==5.4.0